BOCM-358 - WIP refactoring appeals status for inspector endpoints
